pyproject.toml
[project]
name = "local-deployments"
version = "0.1.0"
description = "A Software-Defined Cyber Range manager for Proxmox infrastructure."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "networkx>=3.6.1",
    "proxmoxer>=2.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.14",
]

# --- Tooling Config ---
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# E: pycodestyle violations
# F: pyflakes violations
# I: import-related violations
# B: bugbear violations
# UP: unused imports
select = ["E", "F", "I", "B", "UP"]
ignore = [
    "E501",  # line too long
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = 3.12
strict = true
plugins = ["pydantic.mypy"] # Needed for pydantic models
exclude = ["venv/", ".venv"]

[[tool.mypy.overrides]]
module = "proxmoxer.*"
ignore_missing_imports = true # Proxmoxer has some dynamic imports that mypy can't resolve